--- a/api/cluster/resource/templater.go
+++ b/api/cluster/resource/templater.go
@@ -525,6 +525,7 @@ func createDefaultTransformerEnvVars(modelService *models.Service) models.EnvVar
 	defaultEnvVars = append(defaultEnvVars, models.EnvVar{Name: envModelVersion, Value: modelService.ModelVersion})
 	defaultEnvVars = append(defaultEnvVars, models.EnvVar{Name: envModelFullName, Value: modelService.Name})
 	defaultEnvVars = append(defaultEnvVars, models.EnvVar{Name: envPredictorHost, Value: createPredictorHost(modelService)})
+	defaultEnvVars = append(defaultEnvVars, models.EnvVar{Name: envProtocol, Value: fmt.Sprint(modelService.Protocol)})
 	return defaultEnvVars
 }