--- a/api/cluster/resource/templater.go
+++ b/api/cluster/resource/templater.go
@@ -72,6 +72,7 @@ const (
 	defaultPredictorArtifactLocation = "/mnt/models"
 	defaultHTTPPort                  = 8080
 	defaultGRPCPort                  = 9000
+	defaultPredictorHostPort         = 80
 )
 
 var (
@@ -441,7 +442,7 @@ func createLivenessProbeSpec(path string) *corev1.Probe {
 }
 
 func createPredictorHost(modelService *models.Service) string {
-	return modelService.Name + "-predictor-default." + modelService.Namespace
+	return fmt.Sprintf("%s-predictor-default.%s:%d", modelService.Name, modelService.Namespace, defaultPredictorHostPort)
 }
 
 func createAnnotations(modelService *models.Service, config *config.DeploymentConfig) (map[string]string, error) {