--- a/api/pkg/transformer/server/grpc_server_test.go
+++ b/api/pkg/transformer/server/grpc_server_test.go
@@ -1,0 +1,1 @@
+package server