--- a/api/pkg/transformer/types/pipeline.go
+++ b/api/pkg/transformer/types/pipeline.go
@@ -1,0 +1,8 @@
+package types
+
+type Pipeline string
+
+const (
+	Preprocess  Pipeline = "preprocess"
+	Postprocess Pipeline = "postprocess"
+)