<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the e2e tests
3. Make sure documentation is updated for your PR!

-->

**What this PR does / why we need it**:

Adding GRPC support to standard transformer. This PR doesn't include all the transformation in standard transformer only passthrough to model prediction

## Standard Transformer
* Introduce new types `Payload`. This type will be input or output of standard transformer.
* Change `JSONContainer` into `PayloadObjectContainer` because the output or input of standard transformer doesn't necessary JSON
* Add grpc server implementation including reflection and interceptors 

## Merlin API
* Set `CARAML_PROTOCOL` environment when creating transformer
*  Update transformer simulation so it accepts `Payload`


**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in the block below.
If the PR requires additional action from users switching to the new release, include the string "action required".

For more information about release notes, see kubernetes' guide here: http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note

```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API changes