@pradithya 2022-09-06
Will there be performance degradation to the http protocol when this PR is merged?