@tiopramayudi 2022-09-06
> Will there be performance degradation to the http protocol when this PR is merged?
my hunch said that it is negligible, let me do some benchmark