@tiopramayudi 2022-09-06
> > Will there be performance degradation to the http protocol when this PR is merged?
> 
> my hunch said that it is negligible, let me do some benchmark


I did benchmark using complex standard transformer configuration

old one 
```
old one
goos: darwin
goarch: amd64
pkg: github.com/gojek/merlin/pkg/transformer/server
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_PredictHandler-16    	    2017	    581713 ns/op	  136667 B/op	    1771 allocs/op
PASS
ok  	github.com/gojek/merlin/pkg/transformer/server	2.333s
```

new one
```
goos: darwin
goarch: amd64
pkg: github.com/gojek/merlin/pkg/transformer/server/rest
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_PredictHandler-16    	    3428	    308774 ns/op	  101523 B/op	    1702 allocs/op
PASS
ok  	github.com/gojek/merlin/pkg/transformer/server/rest	2.084s
```

it is faster most likely due to using go 1.18