@pradithya 2022-09-06
> 

Perfect! So around 90% speed improvement and 30% reduced memory usage? That's awesome