@pradithya 2022-09-12
@tiopramayudi It seems that the pyfunc-server picked up the new release in UPI, can you pin this to 0.1.0 to make your PR pass? https://github.com/gojek/merlin/runs/8301655388?check_suite_focus=true